msm: kgsl: Cleanup memdesc flags
We had a ugly mix of flags from usersapce mixed in with internal flags
and the flags were stored in priv and the internal flags were stored
in flags and general confusion all around.

So, external (user) flags moved to memdesc->flags, internal bits moved
to ->priv and now we are ready to handle the full onslaught of user
specified flags.

Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>
Signed-off-by: Rajeev Kulkarni <krajeev@codeaurora.org>

Conflicts:

	drivers/gpu/msm/kgsl_sharedmem.c

Change-Id: If8e88a5260b868b8ec77f26661211d65a5e8de71
Signed-off-by: Srivalli Oguri <oguri@codeaurora.org>