msm: kgsl: Add a flag to context struct for bad timestamp waits
If the userspace calls the waittimestamp ioctl with an invalid
timestamp for a context then set a flag indicating that the ioctl
was called with an invalid ts. If the flag is set then do not
print error message about this ioctl being called with an invalid
timestamp. This is required to prevent the kernel log from
spamming with error messages and causing a watchdog.

CRs-fixed: 374586

Signed-off-by: Shubhraprakash Das <sadas@codeaurora.org>
Signed-off-by: Rajeev Kulkarni <krajeev@codeaurora.org>
(cherry picked from commit 46ee1095077fb629d509944be508fe6da4d8e666)

Change-Id: If90d1d46635dfab09bbaa58e178a17741d13ff21
Signed-off-by: Srivalli Oguri <oguri@codeaurora.org>