[lmk] create a thread for low mem killer instead of using kswapd
https://patchwork.kernel.org/patch/1879571/ by xiaobing.tu@intel.com

Currently low mem killer is registered as a shrinker and will be
invoked by kswapd. Since low mem killer and kswapd has different
criteria to invoke to work -- kswap wakes up when watermark of some
certain order in zone is low, and low mem killer should do its job
when free memory is below the min free threshold. In the case that
the kswapd isn't waked up and lmk is, the killed process selected
by lmk will release memory so that the kswapd doesn't need to work
any more.So it is not appropriate to mix them together. A 20 minutes
average vmstat shows that the iowait/page fault has huge improvement.
In this way, it can reduce the chance of ANR due to high iowait.

orignal:
flt    wa
62.70 2.00

with patch
flt    wa
2.32  0.01

-shenlong: it reduces lagginess when iyokan is in low memory.